OIBSIP-TASK1
A Basic simple calculator is built using html, css and javascript! It takes a little skill to create a calculator with basic mathematical functionalities using CSS, HTML, and JavaScript. Calculators require an interactive interface to perform basic functions such as addition, subtraction, division, and multiplication. I gave a display screen to display the user’s input and give the results. The grid system in CSS is for the alignments of buttons on the calculator. The additional tools i used to build a fully functional calculator includes eventListeners, if-else statements, operators, loops, and so on